Tests added

+++ scheduler.c
@@ -184,57 +184,64 @@ void task_yield()
 */
 void increasePriority(task *t)
 {
+    if(t == NULL){
+        return;
+    }
     // determine what queue the task is in and change its priority to be accurate
     switch (t->taskPriority)
     {
-        case 4:
+    case 4:
+    {
+
+        printf("No higher priority\n");
+        break;
+    }
+    case 3:
+    {
+        addToQueue(t, &p_four);
+        t->taskPriority = 4;
+        for (int i = 0; i < MAX_TASKS; i++)
         {
-            printf("No higher priority\n");
-            break;
-        }
-        case 3:
-        {
-            addToQueue(t, &p_four);
-            for (int i = 0; i < MAX_TASKS; i++)
+            if (p_three[i]->taskid = t->taskid)
             {
-                if (p_three[i]->taskid == t->taskid)
-                {
-                    shift_task(&p_three, i);
-                    break;
-                }
+                shift_task(&p_three, i);
+                break;
             }
-            break;
         }
-        case 2:
+        break;
+    }
+    case 2:
+    {
+        addToQueue(t, &p_three);
+        t->taskPriority = 3;
+        for (int i = 0; i < MAX_TASKS; i++)
         {
-            addToQueue(t, &p_three);
-            for (int i = 0; i < MAX_TASKS; i++)
+            if (p_two[i]->taskid = t->taskid)
             {
-                if (p_two[i]->taskid == t->taskid)
-                {
-                    shift_task(&p_two, i);
-                    break;
-                }
+                shift_task(&p_two, i);
+                break;
             }
-            break;
         }
-        case 1:
+        break;
+    }
+    case 1:
+    {
+        addToQueue(t, &p_two);
+        t->taskPriority = 2;
+        for (int i = 0; i < MAX_TASKS; i++)
         {
-            addToQueue(t, &p_two);
-            for (int i = 0; i < MAX_TASKS; i++)
+            if (p_one[i]->taskid = t->taskid)
             {
-                if (p_one[i]->taskid == t->taskid)
-                {
-                    shift_task(&p_one, i);
-                    break;
-                }
+                shift_task(&p_one, i);
+                break;
             }
-            break;
         }
-        default:
-        {
-            break;
-        }
+        break;
+    }
+    default:
+    {
+        break;
+    }
     }
 }
 
@@ -258,6 +265,7 @@ void deleteFromList(task t)
     {
         if (t.taskid == task_list[i].taskid)
         {
+
             deleteX = i;
             break;
         }
@@ -270,6 +278,7 @@ void deleteFromList(task t)
     task_list[deleteX].taskFunc = NULL;
     task_list[deleteX].taskid = task_list[deleteX + 1].taskid;
     task_list[deleteX].taskPriority = 0;
+    printf("%d %d\n", deleteX, task_list[deleteX].taskPriority);
     task_list[deleteX].context.uc_stack.ss_sp = NULL;
     task_list[deleteX].context.uc_stack.ss_size = 0;
     task_list[deleteX].context.uc_link = NULL;
@@ -320,7 +329,7 @@ void runScheduler()
             {
                 perror("Swap context");
             }
-            if (delete)
+            if (delete && !increase)
             {
                 deleteFromList(*p_three[i]);
                 shift_task(&p_three, i);
@@ -330,6 +339,7 @@ void runScheduler()
             if (increase)
             {
                 increasePriority(p_three[i]);
+                i--;
                 increase = 0;
             }
             i++;
@@ -344,16 +354,19 @@ void runScheduler()
             {
                 perror("Swap context");
             }
-            if (delete)
+            if (delete && !increase)
             {
                 deleteFromList(*p_two[i]);
                 shift_task(&p_two, i);
+
                 i--;
+
                 delete = 0;
             }
             if (increase)
             {
                 increasePriority(p_two[i]);
+                i--;
                 increase = 0;
             }
             i++;
@@ -369,16 +382,18 @@ void runScheduler()
             {
                 perror("Swap context");
             }
-            if (delete)
+            if (delete && !increase)
             {
                 deleteFromList(*p_one[i]);
                 shift_task(&p_one, i);
                 i--;
+
                 delete = 0;
             }
             if (increase)
             {
                 increasePriority(p_one[i]);
+                i--;
                 increase = 0;
             }
             i++;
